+++ misc/harpyenv.yaml
@@ -24,6 +24,7 @@ dependencies:
   - bioconductor-complexheatmap=2.16.0=r43hdfd78af_0
   - bioconductor-iranges=2.34.1=r43ha9d7317_0
   - bioconductor-s4vectors=0.38.1=r43ha9d7317_0
+  - biopython=1.81=py310h2372a71_1
   - boost-cpp=1.78.0=h2c5509c_4
   - boto3=1.28.55=pyhd8ed1ab_0
   - botocore=1.31.55=pyhd8ed1ab_0
@@ -203,10 +204,12 @@ dependencies:
   - pango=1.50.14=ha41ecd1_2
   - parallel=20160622=1
   - paramiko=3.3.1=pyhd8ed1ab_0
+  - pbzip2=1.1.13=h1fcc475_2
   - pcre2=10.40=hc3806b6_0
   - peppy=0.35.7=pyhd8ed1ab_0
   - perl=5.32.1=4_hd590300_perl5
   - perl-threaded=5.32.1=hdfd78af_1
+  - pigz=2.8=h2797004_0
   - pillow=10.0.1=py39h444a776_1
   - pip=23.2.1=py39h06a4308_0
   - pixman=0.42.2=h59595ed_0
@@ -224,6 +227,7 @@ dependencies:
   - pyasn1=0.5.0=pyhd8ed1ab_0
   - pyasn1-modules=0.3.0=pyhd8ed1ab_0
   - pycparser=2.21=pyhd8ed1ab_0
+  - pyfaidx=0.7.2.2=pyhdfd78af_0
   - pygments=2.16.1=pyhd8ed1ab_0
   - pynacl=1.5.0=py39hd1e30aa_3
   - pyopenssl=23.2.0=pyhd8ed1ab_1
@@ -236,10 +240,12 @@ dependencies:
   - python-dateutil=2.8.2=pyhd8ed1ab_0
   - python-fastjsonschema=2.18.0=pyhd8ed1ab_0
   - python-irodsclient=1.1.8=pyhd8ed1ab_0
+  - python-isal=1.5.1=py310h2372a71_0
   - python-tzdata=2023.3=pyhd8ed1ab_0
   - python_abi=3.9=4_cp39
   - pytz=2023.3.post1=pyhd8ed1ab_0
   - pyu2f=0.1.5=pyhd8ed1ab_0
+  - pyvcf3=1.0.3=pyhdfd78af_0
   - pyyaml=6.0.1=py39hd1e30aa_1
   - r-askpass=1.2.0=r43h57805ef_0
   - r-assertthat=0.2.1=r43hc72bb7e_4
@@ -405,9 +411,11 @@ dependencies:
   - vcflib=1.0.9=h146fbdb_4
   - veracitools=0.1.3=py_0
   - wcwidth=0.2.6=pyhd8ed1ab_0
+  - whatshap=2.1=py310h0dbaff4_0
   - wfa2-lib=2.3.4=h4ac6f70_0
   - wheel=0.41.2=py39h06a4308_0
   - wrapt=1.15.0=py39hd1e30aa_1
+  - xopen=1.8.0=py310hff52083_0
   - xorg-kbproto=1.0.7=h7f98852_1002
   - xorg-libice=1.1.1=hd590300_0
   - xorg-libsm=1.2.4=h7391055_0
